+++ Diagnostics/Assets/Scripts/Remote/HTS_Server.cs
@@ -272,7 +272,6 @@ public class HTS_Server : MonoBehaviour
     private void ReceiveFile(string data)
     {
         var parts = data.Split(new char[] { ':' }, 3);
-        for (int k = 0; k < parts.Length; k++) Debug.Log(parts[k]);
         if (parts.Length != 3) return;
         var folder = FileLocations.LocalResourceFolder(parts[0]);
         if (!Directory.Exists(folder))