+++ Diagnostics/Assets/Turandot/Screen/Inputs/Turandot.Screen.CheckListLayout.cs
@@ -12,6 +12,18 @@ namespace Turandot.Screen
     [Serializable]
     public class ChecklistLayout : InputLayout
     {
+        [Category("Behavior")]
+        [Description("If true, does not show Apply button if single selection")]
+        [DisplayName("Auto advance")]
+        public bool AutoAdvance { get; set; }
+
+        [Category("Behavior")]
+        [Description("If true, allow multiple selections")]
+        [DisplayName("Allow multiple")]
+        public bool AllowMultiple { get; set; }
+        private bool ShouldSerializeAllowMultiple() { return false; }
+
+        private bool ShouldSerializeAutoAdvance() { return false; }
         [Category("Design")]
         public string Label { set; get; }
         private bool ShouldSerializeLabel() { return false; }
@@ -24,6 +36,7 @@ namespace Turandot.Screen
         {
             Name = "Checklist";
             Label = "Select one";
+            Items = new BindingList<string>();
         }
 
     }